requirements.txt
requests>=2.31.0
Pillow>=10.0.0
whoosh>=2.7.4
faiss-cpu>=1.7.4
sentence-transformers==2.2.2
huggingface-hub>=0.10.0,<0.16.0
numpy>=1.24.0
pandas>=2.3.0
torch>=2.0.0
nltk>=3.8.1
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-community>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
